feat: Created dynamic manifest config. (#758)
* feat: Created dynamic manifest config.

* fix: Fixed build issues with cloudflare

So long story shorts. This is the issue.
If you use a require(...) or an async import(...) anywhere in the code, the bundler embeded withing wrangler is going to try and look it up, even if it is not within the required files.

Because it encounters imports that are only available to sveltekit, it explodes with a really unclear error.

We cannot disable the bundler, because it needs to lookup between folders inside of .svelte-kit, otherwise it would have been the perfect solution.

For now, at least this works.