fix: nuke time compite. Now takes the smallest last claim time (#1167)
There is an edge case, if you buy a land with very old neighbors. We want to compute the nuke time not based on the neighbor elapsed time since last claim. But since ownership.

I coded a solution for this. Dunno if it works, can't really test it with prod