fix: Fixed lagging events (#1280)
Converted the broadcast channel to a back-pressure enabled mpsc queue.

That means that when catching up with old events, the fetching will not continue until
the land historical task has processed all the new events, ensureing that all events
gets processed.

Note that this might not always be the good solution, it's just that in that case,
it makes sense and is easy to manage, especially due to the fact that on crash, it
will just skip some events (the size of the mpsc queue), and then resume as if nothing
has happenned.

Long term solution would be to store some processing state into the database, where the
tasks pull at their own rate, but we need something that works today :)